+++ Example/SwiftyVIPERExample/Modules/RootViewController.swift
@@ -91,21 +91,6 @@ class RootViewController: UIViewController {
 			make.size.equalTo(self.detailsButton.frame.size)
 		}
     }
-	
-	override func viewWillAppear(_ animated: Bool) {
-		super.viewWillAppear(animated)
-		presenter.viewAppearing()
-	}
-	
-	override func viewDidAppear(_ animated: Bool) {
-		super.viewDidAppear(animated)
-		presenter.viewAppeared()
-	}
-	
-	override func viewDidDisappear(_ animated: Bool) {
-		super.viewDidDisappear(animated)
-		presenter.viewDisappeared()
-	}
 }