+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 
 #### Updated
 - `viewLoaded()` no longer optional for Presenters conforming to `ViewPresenterProtocol`
+
+#### Deprecated
+- `ViewPresenterProtocol.viewAppearing()`
+- `ViewPresenterProtocol.viewAppeared()`
+- `ViewPresenterProtocol.viewDisappeared()`